+++ default-icon-descriptions.js
@@ -79,6 +79,11 @@ export default {
     id: "icon.title.autovexlight",
     comment: "Title for AutovexLight icon"
   },
+  babyonesie: {
+    message: "Baby onesie",
+    id: "icon.title.baby-onesie",
+    comment: "Title for baby onesie icon"
+  },
   backwheeldrive: {
     message: "Rear-wheel drive",
     id: "icon.title.back-wheel-drive",
@@ -128,6 +133,11 @@ export default {
     message: "Bell",
     id: "icon.title.bell",
     comment: "Title for bell icon"
+  },
+  bellfilled: {
+    message: "Bell",
+    id: "icon.title.bell-filled",
+    comment: "Title for bell filled icon"
   },
   bin: {
     message: "Trash can",
@@ -218,6 +228,11 @@ export default {
     message: "Camper",
     id: "icon.title.camping",
     comment: "Title for camping icon"
+  },
+  cancel: {
+    message: "Circle with a diagonal line",
+    id: "icon.title.cancel",
+    comment: "Title for cancel icon"
   },
   car: {
     message: "Car",
@@ -394,10 +409,20 @@ export default {
     id: "icon.title.circle-plus",
     comment: "Title for circle-plus icon"
   },
+  circleplusfilled: {
+    message: "Plus sign inside a circle",
+    id: "icon.title.circle-plus-filled",
+    comment: "Title for circle plus filled icon"
+  },
   circleuser: {
     message: "Circle with user profile",
     id: "icon.title.circle-user",
     comment: "Title for circle-user icon"
+  },
+  circleuserfilled: {
+    message: "Person in a circle",
+    id: "icon.title.circle-user-filled",
+    comment: "Title for circle user filled icon"
   },
   clipboard: {
     message: "Clipboard",
@@ -1074,6 +1099,11 @@ export default {
     id: "icon.title.hourglass",
     comment: "Title for hourglass icon"
   },
+  house: {
+    message: "House",
+    id: "icon.title.house",
+    comment: "Title for house icon"
+  },
   housebed: {
     message: "Bed with a roof over it",
     id: "icon.title.house-bed",
@@ -1083,6 +1113,11 @@ export default {
     message: "Cabin",
     id: "icon.title.house-cabin",
     comment: "Title for house cabin icon"
+  },
+  housefilled: {
+    message: "House",
+    id: "icon.title.house-filled",
+    comment: "Title for house filled icon"
   },
   housemodern: {
     message: "Modern house",
@@ -1389,6 +1424,11 @@ export default {
     id: "icon.title.messages",
     comment: "Title for messages icon"
   },
+  messagesfilled: {
+    message: "Message bubble with smiley face",
+    id: "icon.title.messages-filled",
+    comment: "Title for messages filled icon"
+  },
   microphone: {
     message: "Microphone",
     id: "icon.title.microphone",
@@ -1658,6 +1698,11 @@ export default {
     message: "Phone receiver",
     id: "icon.title.phone",
     comment: "Title for phone icon"
+  },
+  phonebadgecheck: {
+    message: "Smartphone and a badge with a check mark",
+    id: "icon.title.phone-badge-check",
+    comment: "Title for phone badge check icon"
   },
   phonenew: {
     message: "Smartphone",
@@ -2203,6 +2248,11 @@ export default {
     message: "Swap icon",
     id: "icon.title.swap",
     comment: "Title for swap icon"
+  },
+  sweater: {
+    message: "Sweater",
+    id: "icon.title.sweater",
+    comment: "Title for sweater icon"
   },
   swimming: {
     message: "Swimming person",